actualizacion despues de vacaciones

+++ src/main/java/com/grupozeus/telecom/repository/IPersona.java
@@ -4,6 +4,6 @@ import com.grupozeus.telecom.Entitys.Persona;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPersona extends JpaRepository<Persona, Long> {
+public interface IPersona extends JpaRepository<Persona, Integer> {
     
 }
